split up test files
