+++ tools/hangar/tsconfig.json
@@ -2,6 +2,7 @@
   "compilerOptions": {
     "target": "ES2021",
     "moduleResolution": "node",
+    "allowSyntheticDefaultImports": true,
     "strict": true
   }
 }
